virtcontainers: update context id of vsock to uint64
The CID of VSock needs to be change to uint64. Otherwise that leads to
an endianess issue. For more details see
https://github.com/kata-containers/runtime/issues/947

Remove the uint64 introduced by #984

Fixes: #958

+++ virtcontainers/qemu_test.go
@@ -249,14 +249,14 @@ func TestQemuAddDeviceSerialPortDev(t *testing.T) {
 }
 
 func TestQemuAddDeviceKataVSOCK(t *testing.T) {
-	contextID := uint32(3)
+	contextID := uint64(3)
 	port := uint32(1024)
 	vHostFD := os.NewFile(1, "vsock")
 
 	expectedOut := []govmmQemu.Device{
 		govmmQemu.VSOCKDevice{
 			ID:        fmt.Sprintf("vsock-%d", contextID),
-			ContextID: uint64(contextID),
+			ContextID: contextID,
 			VHostFD:   vHostFD,
 		},
 	}
